# Maintain sorted timeslots for performance