… CHANGELOG.md
… qiskit/pulse/utils.py
… test/python/pulse/test_schedule.py
… test/python/pulse/test_timeslots.py