<!--
⚠️ If you do not respect this template, your pull request will be closed.
⚠️ Your pull request title should be short detailed and understandable for all.
⚠️ Also, please add it in the CHANGELOG file under Unreleased section.
⚠️ If your pull request fixes an open issue, please link to the issue.

✅ I have added the tests to cover my changes.
✅ I have updated the documentation accordingly.
✅ I have read the CONTRIBUTING document.
-->

### Summary
Closes #2817. This PR is a follow up to #2484. Performance improves from 29.8s(236s) to 0.217s(0.519s) for the benchmark in #2817.


### Details and comments
This PR improves performance by internally storing Timeslots within TimeslotCollection in a sorted list. This allows O(1) Schedule appends, and other optimizations to be made when creating new TimeslotCollections.

Additional performance plots (different colors are a different number of channels, with same number of instructions appended for each channel).

master:
![image](https://user-images.githubusercontent.com/1553846/61999049-2d624800-b087-11e9-9ac3-817e79afde48.png)
this PR:
![image](https://user-images.githubusercontent.com/1553846/61999051-33f0bf80-b087-11e9-8be8-38e155c0f105.png)